+++ Database/Daos/Bot/BotUserDao.cs
@@ -1,8 +1,8 @@
+using System.Data;
 using WibboEmulator.Database.Interfaces;
 using WibboEmulator.Games.Rooms;
 using WibboEmulator.Games.Rooms.AI;
 using WibboEmulator.Utilities;
-using System.Data;
 
 namespace WibboEmulator.Database.Daos
 {
@@ -37,10 +37,7 @@ namespace WibboEmulator.Database.Daos
             dbClient.RunQuery();
         }
 
-        internal static void UpdatePosition(IQueryAdapter dbClient, int botId, int roomId, int x, int y)
-        {
-            dbClient.RunQuery("UPDATE `bot_user` SET room_id = '" + roomId + "', x = '" + x + "', y = '" + y + "' WHERE id = '" + botId + "'");
-        }
+        internal static void UpdatePosition(IQueryAdapter dbClient, int botId, int roomId, int x, int y) => dbClient.RunQuery("UPDATE `bot_user` SET room_id = '" + roomId + "', x = '" + x + "', y = '" + y + "' WHERE id = '" + botId + "'");
 
         internal static void UpdateLookGender(IQueryAdapter dbClient, int botId, string gender, string look)
         {
@@ -60,15 +57,9 @@ namespace WibboEmulator.Database.Daos
             dbClient.RunQuery();
         }
 
-        internal static void UpdateWalkEnabled(IQueryAdapter dbClient, int botId, bool balkingEnabled)
-        {
-            dbClient.RunQuery("UPDATE `bot_user` SET walk_enabled = '" + WibboEnvironment.BoolToEnum(balkingEnabled) + "' WHERE id = '" + botId + "'");
-        }
+        internal static void UpdateWalkEnabled(IQueryAdapter dbClient, int botId, bool balkingEnabled) => dbClient.RunQuery("UPDATE `bot_user` SET walk_enabled = '" + WibboEnvironment.BoolToEnum(balkingEnabled) + "' WHERE id = '" + botId + "'");
 
-        internal static void UpdateIsDancing(IQueryAdapter dbClient, int botId, bool isDancing)
-        {
-            dbClient.RunQuery("UPDATE `bot_user` SET is_dancing = '" + WibboEnvironment.BoolToEnum(isDancing) + "' WHERE id = '" + botId + "'");
-        }
+        internal static void UpdateIsDancing(IQueryAdapter dbClient, int botId, bool isDancing) => dbClient.RunQuery("UPDATE `bot_user` SET is_dancing = '" + WibboEnvironment.BoolToEnum(isDancing) + "' WHERE id = '" + botId + "'");
 
         internal static void UpdateName(IQueryAdapter dbClient, int botId, string name)
         {
@@ -77,10 +68,7 @@ namespace WibboEmulator.Database.Daos
             dbClient.RunQuery();
         }
 
-        internal static void UpdateRoomBot(IQueryAdapter dbClient, int roomId)
-        {
-            dbClient.RunQuery("UPDATE `bot_user` SET room_id = '0' WHERE room_id = '" + roomId + "'");
-        }
+        internal static void UpdateRoomBot(IQueryAdapter dbClient, int roomId) => dbClient.RunQuery("UPDATE `bot_user` SET room_id = '0' WHERE room_id = '" + roomId + "'");
 
         internal static int InsertAndGetId(IQueryAdapter dbClient, int ownerId, string name, string motto, string figure, string gender)
         {
@@ -102,35 +90,17 @@ namespace WibboEmulator.Database.Daos
             "SELECT '" + userId + "', name, motto, gender, look, '" + roomId + "', walk_enabled, x, y, z, rotation, chat_enabled, chat_text, chat_seconds, is_dancing, is_mixchat FROM `bot_user` WHERE room_id = '" + oldRoomId + "'");
         }
 
-        internal static void UpdateEnable(IQueryAdapter dbClient, int botId, int enableId)
-        {
-            dbClient.RunQuery("UPDATE `bot_user` SET enable = '" + enableId + "' WHERE id = '" + botId + "'");
-        }
+        internal static void UpdateEnable(IQueryAdapter dbClient, int botId, int enableId) => dbClient.RunQuery("UPDATE `bot_user` SET enable = '" + enableId + "' WHERE id = '" + botId + "'");
 
-        internal static void UpdateHanditem(IQueryAdapter dbClient, int botId, int handItem)
-        {
-            dbClient.RunQuery("UPDATE `bot_user` SET handitem = '" + handItem + "' WHERE id = '" + botId + "'");
-        }
+        internal static void UpdateHanditem(IQueryAdapter dbClient, int botId, int handItem) => dbClient.RunQuery("UPDATE `bot_user` SET handitem = '" + handItem + "' WHERE id = '" + botId + "'");
 
-        internal static void UpdateRotation(IQueryAdapter dbClient, int botId, int rotBody)
-        {
-            dbClient.RunQuery("UPDATE `bot_user` SET rotation = '" + rotBody + "' WHERE id = '" + botId + "'");
-        }
+        internal static void UpdateRotation(IQueryAdapter dbClient, int botId, int rotBody) => dbClient.RunQuery("UPDATE `bot_user` SET rotation = '" + rotBody + "' WHERE id = '" + botId + "'");
 
-        internal static void UpdateStatus1(IQueryAdapter dbClient, int botId)
-        {
-            dbClient.RunQuery("UPDATE `bot_user` SET status = '1' WHERE id = '" + botId + "'");
-        }
+        internal static void UpdateStatus1(IQueryAdapter dbClient, int botId) => dbClient.RunQuery("UPDATE `bot_user` SET status = '1' WHERE id = '" + botId + "'");
 
-        internal static void UpdateStatus0(IQueryAdapter dbClient, int botId)
-        {
-            dbClient.RunQuery("UPDATE `bot_user` SET status = '0' WHERE id = '" + botId + "'");
-        }
+        internal static void UpdateStatus0(IQueryAdapter dbClient, int botId) => dbClient.RunQuery("UPDATE `bot_user` SET status = '0' WHERE id = '" + botId + "'");
 
-        internal static void UpdateStatus2(IQueryAdapter dbClient, int botId)
-        {
-            dbClient.RunQuery("UPDATE `bot_user` SET status = '2' WHERE id = '" + botId + "'");
-        }
+        internal static void UpdateStatus2(IQueryAdapter dbClient, int botId) => dbClient.RunQuery("UPDATE `bot_user` SET status = '2' WHERE id = '" + botId + "'");
 
         internal static DataTable GetOneByRoomId(IQueryAdapter dbClient, int roomId)
         {
@@ -139,10 +109,7 @@ namespace WibboEmulator.Database.Daos
             return dbClient.GetTable();
         }
 
-        internal static void Delete(IQueryAdapter dbClient, int userId)
-        {
-            dbClient.RunQuery("DELETE FROM `bot_user` WHERE room_id = '0' AND user_id = '" + userId + "'");
-        }
+        internal static void Delete(IQueryAdapter dbClient, int userId) => dbClient.RunQuery("DELETE FROM `bot_user` WHERE room_id = '0' AND user_id = '" + userId + "'");
 
         internal static DataTable GetAllByUserId(IQueryAdapter dbClient, int userId)
         {
