+++ Database/Daos/User/UserFavoriteDao.cs
@@ -1,24 +1,15 @@
-using WibboEmulator.Database.Interfaces;
 using System.Data;
+using WibboEmulator.Database.Interfaces;
 
 namespace WibboEmulator.Database.Daos
 {
     class UserFavoriteDao
     {
-        internal static void Insert(IQueryAdapter dbClient, int userId, int roomId)
-        {
-            dbClient.RunQuery("INSERT INTO `user_favorite` (`user_id`, `room_id`) VALUES ('" + userId + "','" + roomId + "')");
-        }
+        internal static void Insert(IQueryAdapter dbClient, int userId, int roomId) => dbClient.RunQuery("INSERT INTO `user_favorite` (`user_id`, `room_id`) VALUES ('" + userId + "','" + roomId + "')");
 
-        internal static void Delete(IQueryAdapter dbClient, int userId, int roomId)
-        {
-            dbClient.RunQuery("DELETE FROM `user_favorite` WHERE `user_id` = '" + userId + "' AND `room_id` = '" + roomId + "'");
-        }
+        internal static void Delete(IQueryAdapter dbClient, int userId, int roomId) => dbClient.RunQuery("DELETE FROM `user_favorite` WHERE `user_id` = '" + userId + "' AND `room_id` = '" + roomId + "'");
 
-        internal static void Delete(IQueryAdapter dbClient, int roomId)
-        {
-            dbClient.RunQuery("DELETE FROM `user_favorite` WHERE `room_id` = '" + roomId + "'");
-        }
+        internal static void Delete(IQueryAdapter dbClient, int roomId) => dbClient.RunQuery("DELETE FROM `user_favorite` WHERE `room_id` = '" + roomId + "'");
 
         internal static DataTable GetAll(IQueryAdapter dbClient, int userId)
         {
