+++ Database/Daos/User/UserRoleplayItemDao.cs
@@ -1,14 +1,11 @@
-using WibboEmulator.Database.Interfaces;
 using System.Data;
+using WibboEmulator.Database.Interfaces;
 
 namespace WibboEmulator.Database.Daos
 {
     class UserRoleplayItemDao
     {
-        internal static void Delete(IQueryAdapter dbClient, int userId, int roleplayId)
-        {
-            dbClient.RunQuery("DELETE FROM `user_roleplay_item` WHERE `user_id` = '" + userId + "' AND `rp_id` = '" + roleplayId + "'");
-        }
+        internal static void Delete(IQueryAdapter dbClient, int userId, int roleplayId) => dbClient.RunQuery("DELETE FROM `user_roleplay_item` WHERE `user_id` = '" + userId + "' AND `rp_id` = '" + roleplayId + "'");
 
         internal static DataTable GetAll(IQueryAdapter dbClient, int userId, int roleplayId)
         {
@@ -23,19 +20,10 @@ namespace WibboEmulator.Database.Daos
             return Convert.ToInt32(dbClient.InsertQuery());
         }
 
-        internal static void UpdateAddCount(IQueryAdapter dbClient, int itemId, int count)
-        {
-            dbClient.RunQuery("UPDATE `user_roleplay_item` SET `count` = `count` + '" + count + "' WHERE `id` = '" + itemId + "' LIMIT 1");
-        }
+        internal static void UpdateAddCount(IQueryAdapter dbClient, int itemId, int count) => dbClient.RunQuery("UPDATE `user_roleplay_item` SET `count` = `count` + '" + count + "' WHERE `id` = '" + itemId + "' LIMIT 1");
 
-        internal static void UpdateRemoveCount(IQueryAdapter dbClient, int itemId, int count)
-        {
-            dbClient.RunQuery("UPDATE `user_roleplay_item` SET `count` = `count` - '" + count + "' WHERE `id` = '" + itemId + "' LIMIT 1");
-        }
+        internal static void UpdateRemoveCount(IQueryAdapter dbClient, int itemId, int count) => dbClient.RunQuery("UPDATE `user_roleplay_item` SET `count` = `count` - '" + count + "' WHERE `id` = '" + itemId + "' LIMIT 1");
 
-        internal static void Delete(IQueryAdapter dbClient, int itemId)
-        {
-            dbClient.RunQuery("DELETE FROM `user_roleplay_item` WHERE `id` = '" + itemId + "' LIMIT 1");
-        }
+        internal static void Delete(IQueryAdapter dbClient, int itemId) => dbClient.RunQuery("DELETE FROM `user_roleplay_item` WHERE `id` = '" + itemId + "' LIMIT 1");
     }
 }
